Convert addPerosn to mvvm - Tucker
