+++ ActivityTracker/Views/AddPersonView.swift
@@ -9,15 +9,11 @@ import SwiftUI
 import _PhotosUI_SwiftUI
 
 struct AddPersonView: View {
-    @State private var newName = ""
     
     @Environment(\.managedObjectContext) var moc
     @Environment(\.dismiss) var dismiss
     
-    @State private var personPhotoItem: PhotosPickerItem?
-    @State private var personImageData: Data?
-    @State private var personImage: Image?
-    
+    @State private var viewModel = ViewModel()
     
     var body: some View {
         GeometryReader { geometry in
@@ -25,14 +21,11 @@ struct AddPersonView: View {
                 NavigationStack {
                     Form {
                         Section {
-                            TextField("Name", text: $newName)
+                            TextField("Name", text: $viewModel.newName)
                             if #available(iOS 17.0, *) {
-                                ImagePickerView(photoItem: $personPhotoItem, selectedImageData: $personImageData, imageSize: geometry.size.width * 0.15)
-                                    .onChange(of: personImageData) {
-                                        if let personImageData,
-                                           let uiImage = UIImage(data: personImageData) {
-                                            personImage = Image(uiImage: uiImage)
-                                        }
+                                ImagePickerView(photoItem: $viewModel.personPhotoItem, selectedImageData: $viewModel.personImageData, imageSize: geometry.size.width * 0.15)
+                                    .onChange(of: viewModel.personImageData) {
+                                        viewModel.updatePersonImage()
                                     }
                             }
                         }
@@ -41,31 +34,11 @@ struct AddPersonView: View {
                     .toolbar {
                         ToolbarItem {
                             Button("Save") {
-                                let newPerson = Person(context: moc)
-                                newPerson.id = UUID()
-                                newPerson.name = newName
-                                
-                                if personImage != nil {
-                                    print("has personImage")
-                                    let renderer = ImageRenderer(content: personImage)
-                                    if let uiImage = renderer.uiImage {
-                                        print("has uiImage")
-                                        if let data = uiImage.pngData() {
-                                            let filename = FileManager.getDocumentsDirectory().appendingPathExtension("/personImages").appendingPathComponent("\(newPerson.id!).png")
-                                            print("\(filename)")
-                                            do {
-                                                try data.write(to: filename)
-                                            } catch {
-                                                print("Error writting file: \(error)")
-                                            }
-                                        }
-                                    }
-                                }
-                                
+                                viewModel.savePerson(newPerson: Person(context: moc))
                                 try? moc.save()
                                 dismiss()
                             }
-                            .disabled(newName.isEmpty)
+                            .disabled(viewModel.newName.isEmpty)
                             .padding()
                         }
                     }
